requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-jsonform==2.15.0
django-phonenumber-field==7.0.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dropbox==11.36.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
multidict==6.0.4
numpy==1.23.4
openai==0.27.4
pandas==1.5.0
phonenumberslite==8.13.7
Pillow==9.4.0
ply==3.11
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
tiktoken==0.3.3
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
yarl==1.8.2
